Move popup_function use to style_data method
The actual population of the `_popupContent` attribute based on popup_function is now in the style attribute. This means that the `popup_function` and `popup_attribute` can be populated after the Geojson creation.

The `popup_function` is actually a feature property now to ensure that the `popup_attribute` is set correctly when the function is set after the fact.

Signed-off-by: Henry Walshaw <henry.walshaw@gmail.com>