Require branca 0.3.0, raise error otherwise (#869)
Moving the templates to the class attributes breaks folium when using older branca versions which still have the templates in the init's. The failure is silently, so it may confuse users. So apart from listing the version in requirements.txt I also propose raising an exception when the branca version is older than 0.3.0.