Explicit WMSTileLayer options, accept all **kwargs (#838)
The options for WMSTileLayer were parsed in a function in another module. To make it more explicit and simple I moved this to the class init, just like in the other classes.

Also all **kwargs are now added to the options. This enables users to provide any Leaflet options such as minZoom.