Nested tooltip isInstance testing inside of 'if tooltip' test.
Added additional type testing for kwargs.

If text arg is provided, string conversion is done in the __init__()
constructor instead of the template.

The only kwarg conversion that isn't handled cleanly by
json.dumps(self.kwargs) is 'offset'.
According to https://leafletjs.com/reference-1.0.0.html#tooltip,
It needs to be of a Leaflet 'Point' type. Any ideas on how we could
cleanly prepend 'Point' (or perhaps it needs to be 'L.Point') to our
tuple in the template? Welcome to suggestions.

Progress!