Replace undocumented strftime usage in TimesliderChoropleth example. (#919)
In the TimesliderChoropleth Jupyter notebook example `.strftime('%s')` is used to convert datetime objects to an epoch timestamp string. This is undocumented and system dependent and can lead to unexpected behavior. So as an alternative we can convert the Pandas DateTimeIndex object to epoch timestamps and then to strings, which leads to the same result.