Tooltip classes (simple and geojson/topojson) (#883)
Add two new classes: `Tooltip` and `GeoJsonTooltip`. The first can be used to simply add a fixed text tooltip to most folium objects. The second only works with the `GeoJson` and `TopoJson` classes and can utilize the data fields in those classes to render different tooltip contents for each feature.