Geojson styling overhaul (#1058)
Change the way styling is applied to geojson types. Instead of embedding the styling in the geojson, make a separate mapping that can be applied on the JS side. 

With this change users can supply a file path or a url, set `embed=False` and create a html file that doesn't include the geojson data while still being able to apply a styling function.