added browser print plugin
Happy to build this out further if requested, this is super basic but works, tried to model after your other plugins. A lot more bells and whistles in the js library, https://github.com/Igor-Vladyka/leaflet.browser.print . Also the CDN I add is the first one I found via google, I'll actually be hosting it locally, so noqa on that. might be worth a check, but it looks good.