Fix for issue 1359
This includes a structural change. I added a Class object with
a include method. This follows leaflet's `L.Class.include`
statement. This allows users to override Leaflet class behavior.

The motivating example for this can be found in the added
`test_include` in `test_map.py`. Using an include, users can
override the `createTile` method of `L.TileLayer` and add
a headers.