Issue 1359 (#2159)
* Fix for issue 1359

This includes a structural change. I added a Class object with
a include method. This follows leaflet's `L.Class.include`
statement. This allows users to override Leaflet class behavior.

The motivating example for this can be found in the added
`test_include` in `test_map.py`. Using an include, users can
override the `createTile` method of `L.TileLayer` and add
a headers.

* Close #1359

Add an include statement, that will allow users to override
specific methods at Leaflet level. This allows users to customize
the "createTile" method using a JsCode object.

* Add documentation